A liitle bit more of refactor
